+++ test/integration/property/propertiesSnmp.js
@@ -289,8 +289,8 @@ describe('Snmp', function testSnmp() {
                 },
                 {
                     name: 'destination',
-                    inputValue: ['10.0.10.100', 'fdf5:4153:3300::a', '10.0.10.100'],
-                    expectedValue: ['10.0.10.100', 'fdf5:4153:3300::a', '10.0.10.100']
+                    inputValue: ['10.0.10.100', 'fdf5:4153:3300::a', 'myhost.example.com'],
+                    expectedValue: ['10.0.10.100', 'fdf5:4153:3300::a', 'myhost.example.com']
                 },
                 {
                     name: 'port',