+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Changes to this project are documented in this file. More detail and links can b
 
 ## 1.47.0
 ### Added
-AUTOTOOL-4698: Add support for device DDOS [Example](https://github.com/F5Networks/f5-declarative-onboarding/blob/main/examples/ddos.json)
+- AUTOTOOL-4698: Add support for device DDOS [Example](https://github.com/F5Networks/f5-declarative-onboarding/blob/main/examples/ddos.json)
+- AUTOTOOL-4494: ([GitHub 309](https://github.com/F5Networks/f5-declarative-onboarding/issues/309)) : Unable to set SNMP trap destination to FQDN
 
 ### Fixed